Create random generator class
According to STL's talk <random>, creating a new distribution in line
shouldn't be that bad. That said, the most likely distribution is
going to be [0.0, 1.0], so a uniform_real_distribution for that is
created at class initialization.

If other common ranges come up (i.e. zero to pi), they can also be
initialized with the class, although the fewer the better for
complexity's sake. At some point, I should probably just add a method which
takes some arbitrary distribution, but I don't have any need for that
just yet.