Unsatisfying rework of RandGen propogation through objects
This commit introduces a lot of changes, and I am unsure if they are
for the better. Objects which have their own RandGen now inherit
"PathologicalObject", which provides a clean, consistent, and
deterministic way to create new RandGen for an instance of an
object. Just using the object's RandGen in material sampling, however,
resulted in a massive decrease in performance, as there was a cache
miss for every interaction with the RandGen. To combat this, a new
static thread_local RandGen was added to the sample_hemisphere_uniform
function (which now resides in the RandGen class). This results in
performance on par with the RandGen changes made two commits ago. This
will not scale though, it is completely unreasonable for every
sampling function to have it's own RandGen, and would probably have
serious performance implications.

I am undecided on how I want to resolve this, but with this fix I at
least have a framework to work off of (or to throw out), and I can
work on other features in the meantime. Whatever I come up with, I
look forward to having a better system than this.