Improve region sizing
Each thread renders a region generated at startup. Originally, this
region was computed by getting floor(image_height/nthreads), then
adding any excess to the final image. For example, with an image
height of 99 and 10 threads, the first 9 threads would have a region
of height 9, with the final one a height of 99-(9*9) = 18.

With these changes, the excess is now distributed across all lines, so
in the example above there would now be 9 regions with a height of 10,
and a single region of height 9. This is still not ideal, but it is
much better than before for certain height/thread count combinations.